refactor(todos): change folder structure
Addind containers and components folders
