+++ libs/todos/src/lib/components/controls/controls.component.spec.ts
@@ -6,7 +6,7 @@ import {
   Spectator,
 } from '@ngneat/spectator/jest'
 import { RxState } from '@rx-angular/state'
-import { TodosMainUiStateService } from '../todos-main/todos-main-ui-state.service'
+import { TodosMainUiStateService } from '../../services/todos-main-ui-state.service'
 import { ControlsComponent } from './controls.component'
 
 describe('ControlsComponent', () => {
