+++ libs/todos/src/lib/components/view-todo-entry/view-todo-entry.component.html
