+++ libs/todos/src/lib/routing/todos-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core'
 import { RouterModule, Routes } from '@angular/router'
-import { TodayComponent } from '../today/today.component'
-import { TodosMainComponent } from '../todos-main/todos-main.component'
+import { TodayComponent } from '../containers/today/today.component'
+import { TodosMainComponent } from '../containers/todos-main/todos-main.component'
 
 export const todosRoutes: Routes = [
   {
