+++ libs/todos/src/lib/todos.module.ts
@@ -2,15 +2,15 @@ import { CommonModule } from '@angular/common'
 import { NgModule } from '@angular/core'
 import { FormsModule } from '@angular/forms'
 import { TodosDataModule } from '@parts/todos/data'
-import { AutofocusDirective } from './autofocus.directive'
-import { CheckboxComponent } from './checkbox/checkbox.component'
-import { ControlsComponent } from './controls/controls.component'
-import { NewTodoComponent } from './new-todo/new-todo.component'
+import { CheckboxComponent } from './components/checkbox/checkbox.component'
+import { ControlsComponent } from './components/controls/controls.component'
+import { NewTodoComponent } from './components/new-todo/new-todo.component'
+import { TodoEntryComponent } from './components/todo-entry/todo-entry.component'
+import { ViewTodoEntryComponent } from './components/view-todo-entry/view-todo-entry.component'
+import { TodayComponent } from './containers/today/today.component'
+import { TodosMainComponent } from './containers/todos-main/todos-main.component'
 import { TodosRoutingModule } from './routing/todos-routing.module'
-import { TodayComponent } from './today/today.component'
-import { TodoEntryComponent } from './todo-entry/todo-entry.component'
-import { TodosMainComponent } from './todos-main/todos-main.component'
-import { ViewTodoEntryComponent } from './view-todo-entry/view-todo-entry.component'
+import { AutofocusDirective } from './utils/autofocus.directive'
 
 @NgModule({
   imports: [CommonModule, TodosRoutingModule, FormsModule, TodosDataModule],
